Followed the tutorial up to Chapter 6.
Various features were added before this commit but will be summed up here.

Created:
WelcomeView:
The initial welcome screen for the app.
Includes a welcome message, several tabs to other pages,  and a image.

Also contains a History( )

SuccessView:
When the user completes their workout they will receive  a positive affirmation of a job well done.

 ExerciseView:
Shows what exercise the user is on based on the tab.
Plays an example video of the exercise.
Has a 30 second timer to keep the user focused on their exercise.
Start Exercise and Done button.
Done button moves to the next exercise in the array.
If on final exercise will show the SuccessView( )
Contains rating of the exercise.
Shows a History button that will show HistoryView ( ) when clicked.

 HistoryView:
Show the date and which exercises were performed on those dates.

VideoPlayerView:
Contains the information to play the included exercise videos for the user.

RatingView:
Also the user to select a rating for the exercise on a 1 to 6 scale.

HeaderView:
Shows the current exercise. Also the user to touch the SF symbols to switch to a new tab and exercise.

Standard commits and pushes will occur from now on.